graph: Move StoreError to a separate file
