+++ graph/src/components/store/err.rs
@@ -1,0 +1,95 @@
+use anyhow::{anyhow, Error};
+use thiserror::Error;
+
+use super::{BlockNumber, DeploymentHash};
+use crate::prelude::QueryExecutionError;
+
+#[derive(Error, Debug)]
+pub enum StoreError {
+    #[error("store error: {0}")]
+    Unknown(Error),
+    #[error(
+        "tried to set entity of type `{0}` with ID \"{1}\" but an entity of type `{2}`, \
+         which has an interface in common with `{0}`, exists with the same ID"
+    )]
+    ConflictingId(String, String, String), // (entity, id, conflicting_entity)
+    #[error("unknown field '{0}'")]
+    UnknownField(String),
+    #[error("unknown table '{0}'")]
+    UnknownTable(String),
+    #[error("malformed directive '{0}'")]
+    MalformedDirective(String),
+    #[error("query execution failed: {0}")]
+    QueryExecutionError(String),
+    #[error("invalid identifier: {0}")]
+    InvalidIdentifier(String),
+    #[error(
+        "subgraph `{0}` has already processed block `{1}`; \
+         there are most likely two (or more) nodes indexing this subgraph"
+    )]
+    DuplicateBlockProcessing(DeploymentHash, BlockNumber),
+    /// An internal error where we expected the application logic to enforce
+    /// some constraint, e.g., that subgraph names are unique, but found that
+    /// constraint to not hold
+    #[error("internal constraint violated: {0}")]
+    ConstraintViolation(String),
+    #[error("deployment not found: {0}")]
+    DeploymentNotFound(String),
+    #[error("shard not found: {0} (this usually indicates a misconfiguration)")]
+    UnknownShard(String),
+    #[error("Fulltext search not yet deterministic")]
+    FulltextSearchNonDeterministic,
+    #[error("operation was canceled")]
+    Canceled,
+    #[error("database unavailable")]
+    DatabaseUnavailable,
+    #[error("subgraph forking failed: {0}")]
+    ForkFailure(String),
+}
+
+// Convenience to report a constraint violation
+#[macro_export]
+macro_rules! constraint_violation {
+    ($msg:expr) => {{
+        StoreError::ConstraintViolation(format!("{}", $msg))
+    }};
+    ($fmt:expr, $($arg:tt)*) => {{
+        StoreError::ConstraintViolation(format!($fmt, $($arg)*))
+    }}
+}
+
+impl From<::diesel::result::Error> for StoreError {
+    fn from(e: ::diesel::result::Error) -> Self {
+        StoreError::Unknown(e.into())
+    }
+}
+
+impl From<::diesel::r2d2::PoolError> for StoreError {
+    fn from(e: ::diesel::r2d2::PoolError) -> Self {
+        StoreError::Unknown(e.into())
+    }
+}
+
+impl From<Error> for StoreError {
+    fn from(e: Error) -> Self {
+        StoreError::Unknown(e)
+    }
+}
+
+impl From<serde_json::Error> for StoreError {
+    fn from(e: serde_json::Error) -> Self {
+        StoreError::Unknown(e.into())
+    }
+}
+
+impl From<QueryExecutionError> for StoreError {
+    fn from(e: QueryExecutionError) -> Self {
+        StoreError::QueryExecutionError(e.to_string())
+    }
+}
+
+impl From<std::fmt::Error> for StoreError {
+    fn from(e: std::fmt::Error) -> Self {
+        StoreError::Unknown(anyhow!("{}", e.to_string()))
+    }
+}
